requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
torch>=1.10.0
xgboost>=1.6.0
mlflow>=1.25.0
skorch>=0.12.0
joblib>=1.1.0
PyYAML>=6.0
fastapi>=0.74.0
uvicorn>=0.17.0
pydantic>=1.9.0
